SwiftUI 코드 코디네이터 및 의존성객체 추가하기 (#20)
## :dart: 목적
Presentation 계층에 추가한 SwfitUI 코드와 코디네이터, 의존성 객체(DIContainer) 연결하기
## :pencil:  변경 내용
 > 변경사항을 리스트 형식으로 작성해주세요
 
 - UIkit 과 SwiftUI 네임스페이스 분리
 - 코디네이터 추가
 - DIContainer 추가

## :white_check_mark: 테스트 목록
 > 테스트 목록 작성 후 완료시 체크 해주세요
 - [ ] 
 - [ ] 

## :information_source: 기타
rxSwift 등의 패키지가 모듈화 했을때 경고문구가 발생해서 네임 스페이스로 대체
